Update references to moved plugins
