+++ python/src/aac/plugins/print_spec/__init__.py
@@ -2,9 +2,9 @@
 # NOTE: It is safe to edit this file.
 # This file is only initially generated by aac gen-plugin, and it won't be overwritten if the file already exists.
 
-from aac.plugins import hookimpl
 from aac.cli.aac_command import AacCommand
-from aac.cli.builtin_commands.print_spec.print_spec_impl import print_spec
+from aac.plugins import hookimpl
+from aac.plugins.print_spec.print_spec_impl import print_spec
 
 
 @hookimpl
