+++ python/src/aac/plugins/start_lsp/__init__.py
@@ -2,9 +2,9 @@
 # NOTE: It is safe to edit this file.
 # This file is only initially generated by aac gen-plugin, and it won't be overwritten if the file already exists.
 
-from aac.plugins import hookimpl
 from aac.cli.aac_command import AacCommand, AacCommandArgument
-from aac.cli.builtin_commands.start_lsp.start_lsp_impl import start_lsp
+from aac.plugins import hookimpl
+from aac.plugins.start_lsp.start_lsp_impl import start_lsp
 
 
 @hookimpl
