+++ python/tests/lang/test_server.py
@@ -1,8 +1,6 @@
 from unittest import TestCase
 from unittest.mock import patch
 
-from aac.cli.builtin_commands.start_lsp.start_lsp_impl import start_lsp
-from aac.lang.active_context_lifecycle_manager import get_active_context
 from pygls import uris
 from pygls.lsp import methods
 from pygls.lsp.types import (
@@ -14,6 +12,9 @@ from pygls.lsp.types import (
     InitializeParams,
     Position,
 )
+
+from aac.lang.active_context_lifecycle_manager import get_active_context
+from aac.plugins.start_lsp.start_lsp_impl import start_lsp
 
 from tests.lang.LspTestClient import LspTestClient
 
@@ -32,7 +33,7 @@ class TestLspServer(TestCase):
     def tearDown(self):
         self.client.stop()
 
-    @patch("aac.cli.builtin_commands.start_lsp.start_lsp_impl.server")
+    @patch("aac.plugins.start_lsp.start_lsp_impl.server")
     def test_starts_io_server_when_not_in_dev_mode(self, server):
         result = start_lsp(dev=False)
         self.assertTrue(result.is_success())
